refactor(backend): extract auth routes to dedicated file (#281)
- Move all authentication endpoints from server.ts to routes/auth.ts:
  - POST /auth/register
  - POST /auth/login
  - POST /auth/refresh
  - POST /auth/logout
  - POST /auth/google
  - POST /auth/forgot-password
  - POST /auth/reset-password
  - GET /auth/protected

- Reduce server.ts from 928 lines to 284 lines (~70% reduction)
- Remove duplicate JWT utilities from server.ts
- Register auth routes with prefix '/api/auth'
- No breaking changes to API endpoints

Closes #281
